+++ lib/checkbox_grouped.dart
@@ -1,8 +1,11 @@
 library checkbox_grouped;
 
-export 'src/simple_grouped_checkbox.dart';
-export 'src/custom_grouped_checkbox.dart';
-export 'src/simple_grouped_chips.dart';
-export 'src/simple_grouped_switch.dart';
+export 'src/controller/custom_group_controller.dart';
+export 'src/controller/group_controller.dart';
+export 'src/controller/list_group_controller.dart';
+export 'src/custom_grouped_checkbox.dart' hide CustomGroupedCheckboxState;
+export 'src/list_grouped_checkbox.dart' hide ListGroupedCheckboxState;
 export 'src/show_dialog_grouped_checkbox.dart';
-export 'src/list_grouped_checkbox.dart';
+export 'src/simple_grouped_checkbox.dart' hide SimpleGroupedCheckboxState;
+export 'src/simple_grouped_chips.dart' hide SimpleGroupedChipsState;
+export 'src/simple_grouped_switch.dart' hide SimpleGroupedSwitchState;